More tests, moved classes to separate files
