+++ UnitTests/Sources/NotificationManager/UserNotificationCenterSpy.swift
@@ -1,0 +1,41 @@
+//
+// Copyright 2022 New Vector Ltd
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+import Foundation
+@testable import ElementX
+import UserNotifications
+
+class UserNotificationCenterSpy: UserNotificationCenterProtocol {
+    weak var delegate: UNUserNotificationCenterDelegate?
+    
+    var addRequest: UNNotificationRequest?
+    func add(_ request: UNNotificationRequest, withCompletionHandler completionHandler: ((Error?) -> Void)?) {
+        addRequest = request
+        completionHandler?(nil)
+    }
+    
+    var requestAuthorizationOptions: UNAuthorizationOptions?
+    var requestAuthorizationGrantedReturnValue = false
+    func requestAuthorization(options: UNAuthorizationOptions, completionHandler: @escaping (Bool, Error?) -> Void) {
+        requestAuthorizationOptions = options
+        completionHandler(requestAuthorizationGrantedReturnValue, nil)
+    }
+    
+    var notificationCategoriesValue: Set<UNNotificationCategory>?
+    func setNotificationCategories(_ categories: Set<UNNotificationCategory>) {
+        notificationCategoriesValue = categories
+    }
+}
